% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{check_plink}
\alias{check_plink}
\title{Check if the location for the PLINK is set}
\usage{
check_plink()
}
\value{
TRUE or FALSE
}
\description{
See set_plink() for more information
}
\examples{
check_plink()
}
