% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{get_bcftools_binary}
\alias{get_bcftools_binary}
\title{Find binary for BCFtools}
\usage{
get_bcftools_binary()
}
\value{
Path to BCFtools
}
\description{
Find binary for BCFtools
}
\examples{
if (Sys.info()[["sysname"]] != "Windows") get_bcftools_binary()
}
