% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{set_bcftools}
\alias{set_bcftools}
\title{Set BCFtools binary location}
\usage{
set_bcftools(path = "")
}
\arguments{
\item{path}{If "" (default), then will use genetics.binaRies to get binaries that are appropriate for the detected operating system. Otherwise, provide the path to the BCFtools binary. If NULL then will unset the path.}
}
\value{
NULL or sets the path to the BCFtools binary
}
\description{
Set BCFtools binary location
}
\examples{
# Using the binaries in the genetics.binaRies package.
set_bcftools()
# Setting the binaries to a local path.
\dontrun{
set_bcftools('/path/to/bcftools')
}
# Unset the the path.
set_bcftools(NULL)

}
