% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{set_gcta}
\alias{set_gcta}
\title{Set GCTA binary location}
\usage{
set_gcta(path = "")
}
\arguments{
\item{path}{If "" (default), then will use genetics.binaRies to get binaries that are appropriate for the detected operating system. Otherwise, provide the path to the GCTA binary. If NULL then will unset the path.}
}
\value{
NULL or sets the path to the GCTA binary
}
\description{
Set GCTA binary location
}
\examples{
# Using the binaries in the genetics.binaRies package.
set_gcta() 
# Setting the binaries to a local path.
\dontrun{
set_gcta('/path/to/gcta')
}
# Unset the the path.
set_gcta(NULL) 

}
