% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{set_plink}
\alias{set_plink}
\title{Set PLINK binary location}
\usage{
set_plink(path = "")
}
\arguments{
\item{path}{If "" (default), then will use genetics.binaRies to get binaries that are appropriate for the detected operating system. Otherwise, provide the path to the PLINK binary. If NULL then will unset the path.}
}
\value{
NULL or sets the path to the PLINK binary
}
\description{
Set PLINK binary location
}
\examples{
# Using the binaries in the genetics.binaRies package.
set_plink() 
# Setting the binaries to a local path.
\dontrun{
set_plink('/path/to/plink') 
}
# Unset the the path.
set_plink(NULL) 

}
